Added check of ActivityInfo launch mode when starting activity. When starting a singleTop activity from a singleInstance activity, the onNewIntent() callback was not called when expected. The reason for this was that only the launchMode specified on the intent was checked and not the launchMode specified in the manifest.
Fixes issue 17137.

Johan Viktorsson

Also:
fix NAT detection in SipService
The NAT detection in SipService just checks if the client IP is a
private address. In this case keep alive is enforced. But if you
use a server which also has a private address, keep alive should not
be enabled to save battery.
This patch modifies the NAT detection to also check the server IP.
If both are private, NAT is not detected.

Andre Valentin